Probability (#5)
* Cleaning

Cleaned the files a bit, so the are more clear

* Delete AllFilesinFolder.py

* Further cleaning + adding selected timetrace

* Add files via upload

* Cleaning plotting, does not work yet

* Now the labels do work

* Trying to define groups in the data

Trying to define groups in the data by looking at the mutual distance between two successive points

* Implemented another way to find groups of datapoints

* Update Main.py

* Create REAMDE.txt

* Add files via upload

* Delete REAMDE.txt

* All is working! :D

Everything works, but very slow because there needs to be build a ~5000x2x4000 array by iterating through a for loop, and then the right states have to be found in that array, again by iterating through a double for loop. This must be improved!

* Removed the handmade clusterfinder

Now we should remove some of the for loops

* Add files via upload

* Speeded up the program

* Speeded up the script

* Few little changes

* last version

* Data Handle + Functions

Data handling unobfuscating
Moved things to functions

Plotting does not work for unknown reasons

* Removed Clustering

Removed clustering

* Added pickle

* Add files via upload

* Added Convolution to smoothen Prob. Func.

Added Convolution to smoothen Prob. Func. And plot datapoints in the same state in the same color as the state

* Expanded the colors to timetrace

Expanded the colors to timetrace and plot the position of the peaks in the probability curves

* folder = folder.replace('\\', '\\\\')  :)

* folder = folder.replace('\\', '\\\\')  :D

* Update Functions.py

Fix Ratio