Added some Stuff
States now consist of at least X points using "MinNumOfPoints" in functions.py.
Figures are saved to a separate folder called "Figures"; this gives the ability to save 'Selected'-only and 'All' figures in different folders if we want to.
Smootheness is reduced from 70 to 40. This still has to be optimized.